Update references with refactor API
