+++ src/lib/dataviews/mode/DataViewLocal.ts
@@ -2,7 +2,7 @@ import { Layer } from '@/viz';
 import { AggregationType, aggregate } from '@/data/operations/aggregation/aggregation';
 import { groupValuesByColumn } from '@/data/operations/grouping';
 import { castToNumberOrUndefined } from '@/core/utils/number';
-import { GeoJsonSource } from '@/viz/sources';
+import { GeoJSONSource } from '@/source';
 import { DataViewMode, DataViewData } from './DataViewMode';
 import { CartoDataViewError, dataViewErrorTypes } from '../DataViewError';
 import { CategoryElement } from '../category/CategoryImpl';
@@ -62,7 +62,7 @@ export class DataViewLocal extends DataViewMode {
       return (this.dataSource as Layer).getViewportFeatures(columns);
     }
 
-    return ((this.dataSource as Layer).source as GeoJsonSource).getFeatures(columns);
+    return ((this.dataSource as Layer).source as GeoJSONSource).getFeatures(columns);
   }
 
   private async groupBy(operationColumn: string, operation: AggregationType) {
